+++ smacc_sm_reference_library/sm_dance_bot_strikes_back/servers/service_node_3/src/service_node_3.py
@@ -4,7 +4,6 @@
 import rospy
 import std_srvs
 
-
 if __name__ == "__main__":
 
     def set_bool_service(req):