+++ smacc_client_library/move_base_z_client/custom_planners/backward_local_planner/CHANGELOG.rst
@@ -2,5 +2,5 @@
 Changelog for package backward_local_planner
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
-1.3.2 (2022-11-18)
-------------------
+Forthcoming
+-----------