+++ smacc_client_library/move_base_z_client/custom_planners/undo_path_global_planner/CHANGELOG.rst
@@ -2,7 +2,7 @@
 Changelog for package backward_global_planner
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
-1.3.2 (2022-11-18)
-------------------
+Forthcoming
+-----------
 * Initial SMACC version
 * Contributors: Pablo Iñigo Blasco