+++ smacc_client_library/move_eye/move_eye_client/CHANGELOG.rst
@@ -2,8 +2,8 @@
 Changelog for package move_eye_client
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
-1.3.2 (2022-11-18)
-------------------
+Forthcoming
+-----------
 * move_eye package
 * Contributors: Pabo Iñigo Blasco