+++ smacc_client_library/multirole_sensor_client/CHANGELOG.rst
@@ -2,8 +2,8 @@
 Changelog for package multirole_sensor_client
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
-1.3.2 (2022-11-18)
-------------------
+Forthcoming
+-----------
 
 * Initial SMACC version
 * Contributors: Pablo Iñigo Blasco