+++ smacc_event_generator_library/eg_random_generator/CHANGELOG.rst
@@ -2,8 +2,8 @@
 Changelog for package eg_random_generator
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
-1.3.2 (2022-11-18)
-------------------
+Forthcoming
+-----------
 
 * Renamed sr_conditional & eg_random_generator
 * Contributors: Brett Aldrich