+++ smacc_sm_reference_library/sm_atomic_cb/CHANGELOG.rst
@@ -2,8 +2,8 @@
 Changelog for package sm_atomic_cb
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
-1.3.2 (2022-11-18)
-------------------
+Forthcoming
+-----------
 
 * Merge branch 'master' into melodic-devel
 * adding xterm dependency to examples