+++ smacc_sm_reference_library/sm_calendar_week/CHANGELOG.rst
@@ -2,8 +2,8 @@
 Changelog for package sm_calendar_week
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
-1.3.2 (2022-11-18)
-------------------
+Forthcoming
+-----------
 
 * Initial SMACC version
 * Contributors: Pablo Iñigo Blasco