+++ smacc_sm_reference_library/sm_dance_bot_2/CHANGELOG.rst
@@ -2,7 +2,7 @@
 Changelog for package sm_dance_bot
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
-1.3.2 (2022-11-18)
-------------------
+Forthcoming
+-----------
 * Initial SMACC version
-* Contributors: Pablo Iñigo Blasco, brett2@reelrobotics.com, reelrbtx
+* Contributors: Pablo Iñigo Blasco, brett2@robosoft-ai.com, robosoft-ai