+++ smacc_sm_reference_library/sm_respira_1/CHANGELOG.rst
@@ -2,8 +2,8 @@
 Changelog for package sm_respira_1
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
-1.3.2 (2022-11-18)
-------------------
+Forthcoming
+-----------
 
 * Initial SMACC version
 * Contributors: Pablo Iñigo Blasco