+++ smacc_sm_reference_library/sm_ridgeback_barrel_search_2/CHANGELOG.rst
@@ -2,8 +2,8 @@
 Changelog for package sm_ridgeback_barrel_search_2
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
-1.3.2 (2022-11-18)
-------------------
+Forthcoming
+-----------
 
 * Added gazebo camera changes to sm_opencv and sm_ridgeback_barrel_search_2
 * Merge branch 'master' into melodic-devel