+++ smacc_sm_reference_library/sm_three_some/CHANGELOG.rst
@@ -2,8 +2,8 @@
 Changelog for package sm_three_some
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
-1.3.2 (2022-11-18)
-------------------
+Forthcoming
+-----------
 
 * Initial SMACC version
 * Contributors: Pablo Iñigo Blasco