+++ smacc_state_reactor_library/sr_all_events_go/CHANGELOG.rst
@@ -2,8 +2,8 @@
 Changelog for package sr_all_events_go
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
-1.3.2 (2022-11-18)
-------------------
+Forthcoming
+-----------
 
 * Initial SMACC version
 * Contributors: Pablo Iñigo Blasco