+++ smacc_state_reactor_library/sr_event_countdown/CHANGELOG.rst
@@ -2,8 +2,8 @@
 Changelog for package sr_event_countdown
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
-1.3.2 (2022-11-18)
-------------------
+Forthcoming
+-----------
 
 * Renamed event_countdown to sr_event_countdown
 * Contributors: Brett Aldrich